requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
idna==3.4
python-dotenv==1.0.0
python-gitlab==3.15.0
pytz==2023.3.post1
requests==2.31.0
requests-toolbelt==1.0.0
urllib3==2.0.7
